Use integrations in benchmarks
